replaced window.lua with the separate-view version
